Bug fix for S-meter max value
Swapped logic for max-range-check and second-to-max-range-check in both (if/else, Non-FM/FM) branches in uint8_t getStrength() function. Existing logic will never return max-range value, always returning second-to-max-range value. This fix swaps the last two lines of each of the "if{}" and "else{}" branches to compare/return the max-range-value immediately before checking/returning the second-to-max-range-value.